# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Article API",
        "version": "0.1"
    },
    "paths": {
        "/api/articles/search": {
            "get": {
                "tags": [
                    "article"
                ],
                "summary": "List of articles",
                "description": "List of articles",
                "operationId": "articleList",
                "parameters": [
                    {
                        "name": "q",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "NFT"
                    },
                    {
                        "name": "from_date",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2021-11-02"
                    },
                    {
                        "name": "to_date",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2021-11-02"
                    },
                    {
                        "name": "typename",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "news"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#Article"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    }
}